refactor: 유저, 방 도메인 패키지로 이동
