+++ src/main/java/chess/controller/user/UserController.java
@@ -3,8 +3,8 @@ package chess.controller.user;
 import chess.controller.CommandMapper;
 import chess.controller.Controller;
 import chess.controller.session.RoomSession;
-import chess.controller.session.User;
 import chess.controller.session.UserSession;
+import chess.domain.user.User;
 import chess.dto.NameDto;
 import chess.dto.UserDto;
 import chess.service.UserService;
