+++ src/main/java/chess/domain/user/User.java
@@ -1,4 +1,4 @@
-package chess.controller.session;
+package chess.domain.user;
 
 public class User {
     private final int id;
