+++ src/test/java/chess/domain/room/RoomTest.java
@@ -1,0 +1,21 @@
+package chess.domain.room;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.jupiter.api.DisplayNameGeneration;
+import org.junit.jupiter.api.DisplayNameGenerator;
+import org.junit.jupiter.api.Test;
+
+@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
+@SuppressWarnings("NonAsciiCharacters")
+class RoomTest {
+
+    @Test
+    void 방이_정상_생성된다() {
+        // given
+        final Room room = new Room(1, "체스 한 게임?");
+
+        // expect
+        assertThat(room.getName()).isEqualTo("체스 한 게임?");
+    }
+}
